updated endpoints

+++ helpers.py
@@ -5,6 +5,7 @@ import sys
 from collections import defaultdict
 
 import numpy as np
+import cv2
 from PIL import Image
 
 
@@ -71,10 +72,16 @@ def hash_files(location):
         ext = f.split('.')[-1]
         link_src = f
         link_dest = os.path.join('/app/static', name)
-        os.symlink(link_src, link_dest)
+
+        if not os.path.exists(link_dest):
+            os.symlink(link_src, link_dest)
+
         just_hash_name = link_dest.split('_')[-1]
-        just_hash_name = os.path.join('/app/static',just_hash_name)
-        os.symlink(link_src, just_hash_name)
+        just_hash_name = os.path.join('/app/static', just_hash_name)
+
+        if not os.path.exists(just_hash_name):
+            os.symlink(link_src, just_hash_name)
+
         sys.stdout.flush()
         if ext in ['avi', 'mov', 'mp4']:
             file_hash = file_digest(f)
@@ -82,6 +89,9 @@ def hash_files(location):
             ret_val[file_hash]['Hash'] = file_hash
             ret_val[file_hash]['file_content_hash'] = file_hash
             ret_val[file_hash]['Name'] = name
+            video_file = cv2.VideoCapture(f)
+            video_length = int(video_file.get(cv2.CAP_PROP_FRAME_COUNT))
+            ret_val[file_hash]['frames'] = video_length
             print('found: {0}:{1}'.format(file_hash, f.split('/')[-1]))
             sys.stdout.flush()
     return ret_val
